Separate SynchronizedImageReader
